Tidy cron module up
